putMeeple(x,y,px,py) is working now. Added a few other things too.
Please see the Java comments I left for clarification about how
putMeeple works.

Feel free to write more meeple tests but I am fairly certain that I
accounted for all possibilities.

I changed the data structure to reflect proper Cartesian coordinates in
Java.  tiles should be (x,y) where x increases from left to right and y
increases from top to bottom. Everything seems to be working but please
remember to account for this correction in the UI.

I also added in a state tracking method called get_playersTurn() which
returns as an int which players turn it is. (from 1 to _playerNumber)

additionally there is a method that returns void but increments the turn
counter. this should be called whenever someones turn is finished.

putMeeple decrements the players supply whenever it returns true.